feat(contracts): move shared CritiqueRoundSummary / CritiqueRunStatus into the contracts package
The daemon's persistence layer was the source of truth for three pieces
of public critique state: the CritiqueRoundSummary record, the terminal
CritiqueRunStatus union, and the CRITIQUE_RUN_STATUSES enumeration that
defines the canonical display order. None of those are daemon-private:
the web layer (Phase 7 reducer + Theater UI) needs the same shapes to
render run history, status filters, and the rerun affordance, and the
root AGENTS.md requires shared web/daemon DTOs to live in
packages/contracts so the two sides can't drift.

Move all three to packages/contracts/src/critique.ts and have
apps/daemon/src/critique/persistence.ts re-export them so existing
daemon imports (orchestrator, server, run-registry, tests) keep
working unchanged. The CHECK-constraint-aware ALL_VALID_STATUSES set
that adds the in-flight 'running' value stays daemon-side because it
is strictly a DB invariant, not a public surface.

This unblocks the web layer's eventual rerun / history wiring without
forcing it to either duplicate the shapes or import daemon source
(which would violate the apps/web → apps/daemon boundary).
